Switching to use ThreadUtil.run() instead of ExecutorService. For some reason all of the threads block and one CPU is pegged at 100%. Removed unecessary synchonized decorators as well
git-svn-id: https://database.cs.brown.edu/svn/hstore/branches/markov-branch@1845 7a8f3d5b-8c5f-44cb-9125-2414a256df87

+++ src/frontend/edu/brown/markov/Edge.java
@@ -27,7 +27,7 @@ public class Edge extends AbstractEdge implements Comparable<Edge> {
     /**
      * This is the probability that the source of the edge will transition to the destination vertex
      */
-    public double probability;
+    public float probability;
 
     /**
      * This is the total number of times that we have traversed over this edge
@@ -55,7 +55,7 @@ public class Edge extends AbstractEdge implements Comparable<Edge> {
     public Edge(IGraph<Vertex, Edge> graph, int hits, double probability) {
         super(graph);
         this.hits = hits;
-        this.probability = probability;
+        this.probability = (float)probability;
     }
 
     @Override
